fix(lifts): write default state to local storage
Fixes a bug where a new user can navigate to app, select a lift,
and refresh the page. Before the lift would no longer exist because a
new uuid was generated. The state was only saved if changes were made to
the default. Now the state is saved right away.